Update more int->NSInteger for window numbers
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/gnustep_testplant_branch@40349 72102866-910b-0410-8b05-ffd578937521

+++ Source/win32/WIN32Server.m
@@ -776,8 +776,8 @@ LRESULT CALLBACK windowEnumCallback(HWND hwnd, LPARAM lParam)
 }
 
 - (int) findWindowAt: (NSPoint)screenLocation 
-           windowRef: (int*)windowRef 
-           excluding: (int)win
+           windowRef: (NSInteger*)windowRef
+           excluding: (NSInteger)win
 {
   HWND hwnd;
   POINT p;
